move AppComponent/AppContainer to templates/ and pages/ directory
