+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import AppContainer from "./AppContainer";
+import AppContainer from "./pages/AppContainer";
 
 ReactDOM.render(<AppContainer />, document.getElementById("root"));
